Beginstudentportal

+++ classes/Token.php
@@ -19,13 +19,15 @@ class Token
 	}
 
 
-	public static function check($token, $token_name = null): bool
+	public static function check($token, $token_name = null,$retain_token=false): bool
 	{
 		if (!isset($token_name)) {
 			$token_name = Config::get('session/token_name');
 		}
 		if (Session::exists($token_name) && $token === Session::get($token_name)) {
-			Session::delete($token_name);
+			if(!$retain_token){
+				Session::delete($token_name);
+			}
 			return true;
 		}
 		return false;
